requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-database-url==3.0.1
python-dotenv==1.1.1
psycopg2-binary==2.9.11
firebase-admin==7.1.0
reportlab==4.0.9
requests>=2.31.0
webauthn>=2.7.1
ultralytics>=8.0.0
